Fix bug #2481237: Maps directory file chooser should look for directory
This adds a parameter to vik_file_entry_new() to specify if we are looking for
a file or a directory. There is also a new parameter for widget type:
VIK_LAYER_WIDGET_FOLDERENTRY in addition to VIK_LAYER_WIDGET_FILEENTRY. Now, we
can correctly select a directory when looking at the maps directory property of
maps layers.

Thanks Andrew Lunn to have reported the bug.